Modularized serializer
