+++ activities/serializer/model_serializers.py
@@ -1,0 +1,64 @@
+"""Module for serializing data before respond a request."""
+from typing import Any
+from rest_framework import serializers, exceptions
+from .. import models
+from . import custom_validator
+
+
+class ActivitiesSerializer(serializers.ModelSerializer):
+    """Serialized activity."""
+
+    people = serializers.ReadOnlyField()
+    can_join = serializers.ReadOnlyField()
+    host = serializers.SerializerMethodField()
+
+    class Meta:
+        """Activity serializer META class."""
+
+        model = models.Activity
+        fields = ('__all__')
+
+    def get_host(self, obj: models.Activity) -> list[Any]:
+        """Return list of activity host.
+
+        :param obj: Instance of activity model
+        :return: List of activity host id
+        """
+        act_host = models.Attend.objects.filter(activity=obj, is_host=True)
+        host_ids = [attend.user_id for attend in act_host]
+
+        return host_ids
+
+
+class AttendSerializer(serializers.ModelSerializer):
+    """Serialized Attend."""
+
+    class Meta:
+        """Attend serializer META class."""
+
+        model = models.Attend
+        fields = ('__all__')
+
+        validators = [
+            custom_validator.CustomMsgUniqueTogetherValidator(
+                queryset=model.objects.all(),
+                fields=['user', 'activity']
+            ),
+            custom_validator.CanJoinValidator()
+        ]
+
+    def get_attend(self, activity_id: int, user_id: int) -> Any:
+        """Return attend object specify by activity_id and user_id.
+
+        :param activity_id: Id of activity model instance
+        :param user_id: Id of user model instance
+        :raises exceptions.APIException: if Attend object with specify id not exist.
+        :return: None
+        """
+        try:
+            return models.Attend.objects.get(
+                activity__id=activity_id,
+                user__id=user_id
+            )
+        except models.Attend.DoesNotExist:
+            raise exceptions.APIException("You've never join this activity")
